requirements.txt
pandas>=2.0
yfinance>=0.2.30
pyyaml>=6.0